requirements.txt
PyQt5==5.15.11
google-generativeai==0.8.4
Pillow==11.1.0
reportlab==4.3.1
python-dotenv==1.0.0 